refactor(cisco_ios): consolidate dir parser into iosxe sibling
Removes the duplicate src/muninn/parsers/ios/dir.py and instead stacks
CISCO_IOS register decorators onto the existing IOS-XE DirParser, which
now also handles the narrow-terminal filename-wrap case used by IOS
captures (e.g. Catalyst 3750-X).
